+++ src/event-custom/HISTORY.md
@@ -1,6 +1,12 @@
 Custom Event Infrastructure Change History
 ==========================================
 
+@VERSION@
+------
+
+* Fixed regression introduced in 3.10.0, where `EventTarget.detach('cat|*')` 
+  would throw an exception, when the EventTarget was configured with a prefix.
+
 3.11.0
 ------